NATIVE-306: Refactor withCustomNewsProvider to avoid function usage
