+++ src/modules/app/__tests__/createNavigateToNews.spec.js
@@ -2,7 +2,7 @@
 
 import createNavigationScreenPropMock from '../../test-utils/createNavigationScreenPropMock'
 import createNavigateToNews from '../createNavigateToNews'
-import { LOCAL } from '../../../routes/news/containers/WithCustomNewsProvider'
+import { LOCAL } from '../../../routes/news/hocs/withCustomNewsProvider'
 
 describe('createNavigateToNews', () => {
   it('should generate key if not supplied with at least 6 chars and use it for both navigation and redux actions', () => {
