+++ src/modules/endpoint/sagas/watchFetchNews.js
@@ -14,7 +14,7 @@ import type { DataContainer } from '../DataContainer'
 import { ContentLoadCriterion } from '../ContentLoadCriterion'
 import isPeekingRoute from '../selectors/isPeekingRoute'
 import ErrorCodes, { fromError } from '../../error/ErrorCodes'
-import { LOCAL } from '../../../routes/news/containers/WithCustomNewsProvider'
+import { LOCAL } from '../../../routes/news/hocs/withCustomNewsProvider'
 import loadLocalNews from './loadLocalNews'
 import loadTunews from './loadTunews'
 import loadLanguages from './loadLanguages'
