+++ src/routes/news/containers/NewsContainer.js
@@ -19,8 +19,8 @@ import createNavigateToNews from '../../../modules/app/createNavigateToNews'
 import type { Dispatch } from 'redux'
 
 import type { NavigationScreenProp } from 'react-navigation'
-import withCustomPayloadProvider, { TUNEWS } from './WithCustomNewsProvider'
-import type { StatusPropsType } from './WithCustomNewsProvider'
+import withCustomPayloadProvider, { TUNEWS } from '../hocs/withCustomNewsProvider'
+import type { StatusPropsType } from '../hocs/withCustomNewsProvider'
 import { CityModel } from '@integreat-app/integreat-api-client'
 import * as React from 'react'
 import { mapProps } from 'recompose'
